作品完善

+++ src/main/java/com/ecommerce/demo/entity/model/OrderInfo.java
@@ -26,7 +26,7 @@ public class OrderInfo {
     /**
      * 订单成交价
      */
-    private double sellPrice;
+    private float sellPrice;
 
     /**
      * 订单状态
